% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_clinical}
\alias{check_clinical}
\title{check meta clinical}
\usage{
check_clinical(clinical_dat, clinical_meta, datatype)
}
\arguments{
\item{clinical_dat}{Dataframe with Clinical Data.}

\item{clinical_meta}{Dataframe indicating the type of data.}

\item{datatype}{One of \code{PATIENT_ATTRIBUTES} or \code{SAMPLE_ATTRIBUTES}.}
}
\value{
stop if requirement are not met
}
\description{
Verify  dataframes used for \code{create_meta_clinical}
}
