% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_expr}
\alias{check_expr}
\title{check expr data}
\usage{
check_expr(expr_matrix)
}
\arguments{
\item{expr_matrix}{Dataframe of expression.}
}
\value{
stop if requirement are not met
}
\description{
Verify expression data
}
