% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counts}
\alias{counts}
\title{counts}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 100 rows and 28 columns.
}
\usage{
counts
}
\description{
Am example of count matrix with 100 genes and 28 samples. Counts have been generated randomly.
}
\keyword{datasets}
