% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_cancer_type}
\alias{create_cancer_type}
\title{create cancer type}
\usage{
create_cancer_type(
  folder_path,
  type_of_cancer,
  name,
  dedicated_color,
  parent_type_of_cancer
)
}
\arguments{
\item{folder_path}{Path of folder where to save files.}

\item{type_of_cancer}{The cancer type abbreviation, e.g., "brca".}

\item{name}{The name of the cancer type, e.g., "Breast Invasive Carcinoma".}

\item{dedicated_color}{CSS color name of the color associated with this cancer study, e.g., "HotPink". See this href{https://www.w3.org/TR/css-color-3/#svg-color}{list}for supported names,
and follow the href{https://en.wikipedia.org/wiki/List_of_awareness_ribbons}{awareness ribbons} color schema. This color is associated with the cancer study on various web pages within
the cBioPortal.}

\item{parent_type_of_cancer}{The type_of_cancer field of the cancer type of which this is a subtype, e.g., "Breast". ℹ️ : you can set parent to tissue, which is the reserved word to place the
given cancer type at "root" level in the "studies oncotree" that will be generated in the homepage (aka query page) of the portal.}
}
\value{
Write 2 tab delimited file in \code{folder_path}.
}
\description{
cBioPortal Doc indicates that "If the type_of_cancer specified in the \code{meta_study.txt} does not yet exist in the type_of_cancer database table,
a meta_cancer_type.txt file is also mandatory." The function creates the meta file \code{meta_cancer_type.txt} and the data file associated with it \code{cancer_type.txt}.
}
\seealso{
https://docs.cbioportal.org/file-formats/#cancer-study
}
