% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_clinical}
\alias{create_clinical}
\title{create  clinical files}
\usage{
create_clinical(
  folder_path,
  genetic_alteration_type = "CLINICAL",
  cancer_study_identifier,
  datatype,
  clinical_dat,
  clinical_meta
)
}
\arguments{
\item{folder_path}{Path of folder where to save files.}

\item{genetic_alteration_type}{CLINICAL.}

\item{cancer_study_identifier}{Same value specified in \code{meta_study.txt}.}

\item{datatype}{one of \code{PATIENT_ATTRIBUTES} or \code{SAMPLE_ATTRIBUTES}.}

\item{clinical_dat}{A dataframe with first column \code{PATIENT_ID} or \code{SAMPLE_ID} and other columns of interest.
.  and a column \code{datatype} with values \code{STRING}, \code{NUMBER} or \code{BOOLEAN}, a numeric column called \code{attr_priority} and a an upper case name ot the attribute
.  in column \code{attr_name}.}

\item{clinical_meta}{A dataframe with a column \code{var_int} containing \code{colnames(clinical_dat)}, a column \code{description} with description of variable.}
}
\value{
Write 2 tab delimited file in \code{folder_path}.
}
\description{
Create clinical files \code{meta_clinical_sample.txt} or \code{meta_clinical_patient.txt} and relative data files \code{data_clinical_patient.txt} or
\code{data_clinical_sample.txt}
}
\details{
See Note on survival plots in https://docs.cbioportal.org/file-formats/#cancer-study for additional details on
survival plots.
}
\seealso{
https://docs.cbioportal.org/file-formats/#cancer-study
}
