% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_expression}
\alias{create_expression}
\title{create expression files}
\usage{
create_expression(
  folder_path,
  cancer_study_identifier,
  genetic_alteration_type = "MRNA_EXPRESSION",
  datatype,
  stable_id,
  source_stable_id = NULL,
  show_profile_in_analysis_tab = FALSE,
  profile_name,
  profile_description,
  expr_matrix,
  gene_id_type,
  gene_panel = NULL
)
}
\arguments{
\item{folder_path}{Path of folder where to save files.}

\item{cancer_study_identifier}{Same value specified in \code{meta_study.txt}.}

\item{genetic_alteration_type}{MRNA_EXPRESSION.}

\item{datatype}{one of \code{CONTINUOUS}, \code{DISCRETE} or \code{Z-SCORE}.}

\item{stable_id}{Check \verb{Supported stable_id values for MRNA_EXPRESSION} in https://docs.cbioportal.org/file-formats/#expression-data.}

\item{source_stable_id}{Required when both conditions are true: (1) datatype = Z-SCORE and (2)
this study contains GSVA data. Should contain stable_id of the expression file for which this Z-SCORE file is the statistic.}

\item{show_profile_in_analysis_tab}{false (you can set to true if Z-SCORE to enable it in the oncoprint, for example).}

\item{profile_name}{A name for the expression data, e.g., "mRNA expression (microarray)".}

\item{profile_description}{A description of the expression data, e.g., "Expression levels (Agilent microarray).".}

\item{expr_matrix}{A matrix with expression data.}

\item{gene_id_type}{Either "hugo" or "entrez".}

\item{gene_panel}{Optional gene panel stable id.}
}
\value{
Write 2 tab delimited files in \code{folder_path}.
}
\description{
Create expression files \code{meta_clinical_sample.txt} and relative data files \code{data_expression.txt}.
}
\seealso{
\url{https://docs.cbioportal.org/file-formats/#expression-data}
}
