% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_meta_study}
\alias{create_meta_study}
\title{create meta study}
\usage{
create_meta_study(
  folder_path = getwd(),
  type_of_cancer,
  cancer_study_identifier,
  name,
  description,
  citation,
  pmid,
  groups,
  add_global_case_list = "false",
  tags_file,
  reference_genome
)
}
\arguments{
\item{folder_path}{Path of folder where to save \code{meta_study.txt}.}

\item{type_of_cancer}{The cancer type abbreviation, e.g., "brca". This should be the same cancer type as specified in the meta_cancer_type.txt file, if available. The type can be "mixed" for studies with multiple cancer types. Check href{https://gdc.cancer.gov/resources-tcga-users/tcga-code-tables/tcga-study-abbreviations}{link} for list of available.}

\item{cancer_study_identifier}{A string used to uniquely identify this cancer study within the database, e.g., "brca_joneslab_2013".}

\item{name}{The name of the cancer study, e.g., "Breast Cancer (Jones Lab 2013)".}

\item{description}{A description of the cancer study, e.g., "Comprehensive profiling of 103 breast cancer samples. Generated by the Jones Lab 2013". This description may contain one or more URLs to relevant information.}

\item{citation}{A relevant citation, e.g., "TCGA, Nature 2012".}

\item{pmid}{One or more relevant pubmed ids (comma separated without whitespace). If used, the field citation has to be filled, too.}

\item{groups}{When using an authenticating cBioPortal, lists the user-groups that are allowed access to this study. Multiple groups are separated with a semicolon ";". The study will be invisible to users not in at least one of the listed groups, as if it wasn't loaded at all. e.g., "PUBLIC;GDAC;SU2C-PI3K". see User-Authorization for more information on groups.}

\item{add_global_case_list}{set to 'true' if you would like the "All samples" case list to be generated automatically for you. See also Case lists.}

\item{tags_file}{The file name containing custom study tags for the study tags.}

\item{reference_genome}{The study reference genome (e.g. hg19, hg38). Without specifying this property, the study will be assigned to the reference genome specified in portal.properties (property ucsc.build).}
}
\value{
Write a tab delimited file in \code{folder_path}.
}
\description{
create meta study
}
\seealso{
https://docs.cbioportal.org/file-formats/#cancer-study
}
