% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_patients_datatype}
\alias{df_patients_datatype}
\title{df_patients_datatype}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 5 columns.
}
\usage{
df_patients_datatype
}
\description{
An example dataframe of the sample datatype to be used
}
\keyword{datasets}
