% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_samples}
\alias{df_samples}
\title{df_samples}
\format{
An object of class \code{data.frame} with 28 rows and 7 columns.
}
\usage{
df_samples
}
\description{
An example dataframe of sample info related to the \code{counts} matrix.
}
\keyword{datasets}
