% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_samples_datatype}
\alias{df_samples_datatype}
\title{df_samples_datatype}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 5 columns.
}
\usage{
df_samples_datatype
}
\description{
An example dataframe of sample info datatype to be used
}
\keyword{datasets}
